requirements.txt
-i https://pypi.org/simple
astroid==2.11.6
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
jinja2==3.1.2
lazy-object-proxy==1.7.1
markupsafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pycryptodome==3.15.0
pylint==2.14.4
python-dotenv==0.20.0
requests==2.28.1
setuptools==63.1.0
tinytuya==1.6.1
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.3.0
urllib3==1.26.9
werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
